Refactored evidence component, seperated evidence top bar
